# docs/swagger.yaml
basePath: /
definitions:
...
  server.refreshTokenReq:
    properties:
      userId:
        type: integer
    type: object
  server.refreshTokenRes:
    properties:
      accessToken:
        type: string
    type: object
  server.resendVerificationOTPReq:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  server.verifyAccountReq:
    properties:
      email:
        type: string
      otp:
        type: string
    required:
    - email
    - otp
    type: object
  utils.APIError:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
host: 127.0.0.1:8080
info:
  contact:
    email: ali.jabar@refinesoftware.net
    name: Ali Jabar
  description: Afrad is a RESTful API for a clothing e-commerce platform.
  license:
    name: MIT License
    url: https://img.shields.io/badge/License-MIT-yellow.svg
  title: Afrad API
  version: "1.0"
paths:
...
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Rotates a valid refresh token and returns a new access token. Requires
        refresh token in cookie.
      parameters:
      - description: User ID
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/server.refreshTokenReq'
      produces:
      - application/json
      responses:
        "200":
          description: New access token
          schema:
            $ref: '#/definitions/server.refreshTokenRes'
        "400":
          description: Bad request or missing refresh_token cookie
          schema:
            $ref: '#/definitions/utils.APIError'
        "401":
          description: Invalid, expired, or revoked session
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.APIError'
      security:
      - RefreshTokenCookie: []
      summary: Refresh Tokens
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - multipart/form-data
      description: Registers a new user with optional profile image and sends a verification
        OTP via email.
      parameters:
      - description: First Name
        in: formData
        name: firstName
        required: true
        type: string
      - description: Last Name
        in: formData
        name: lastName
        required: true
        type: string
      - description: Email
        in: formData
        name: email
        required: true
        type: string
      - description: Phone Number
        in: formData
        name: phoneNumber
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      - description: Optional Profile Image
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: user created
          schema:
            type: string
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Register User
      tags:
      - Auth
  /auth/resend-verification:
    post:
      consumes:
      - application/json
      description: Resends an OTP code to a user's email if the account is not yet
        verified.
      parameters:
      - description: Email for which to resend OTP
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/server.resendVerificationOTPReq'
      produces:
      - application/json
      responses:
        "200":
          description: check your email for otp
          schema:
            type: string
        "400":
          description: Bad request, invalid input, or already verified
          schema:
            $ref: '#/definitions/utils.APIError'
        "403":
          description: OTP request limit reached
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Resend Verification OTP
      tags:
      - Auth
  /auth/verify-account:
    post:
      consumes:
      - application/json
      description: Verifies a user's account using email and OTP.
      parameters:
      - description: Verification Data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/server.verifyAccountReq'
      produces:
      - application/json
      responses:
        "200":
          description: your account has been verified
          schema:
            type: string
        "400":
          description: Bad request or invalid OTP
          schema:
            $ref: '#/definitions/utils.APIError'
        "401":
          description: OTP expired
          schema:
            $ref: '#/definitions/utils.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.APIError'
      summary: Verify Account
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {your_jwt_token}" to authenticate.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"